requirements.txt
requests==2.24.0
flask==1.1.2
PyYAML==5.3.1
psycopg2==2.7.7
SQLAlchemy==1.3.18